requirements.txt
streamlit>=1.40.0
openai>=1.0.0
python-dotenv>=1.0.0
st-paywall
httpx-oauth>=0.13.0
stripe